Get rid of Behavior interfaces. Fixes #5
Now all Elements have hardcoded all behavior methods and properties
so as the user does not care about the js hierarchy which is confusing
in java.
Removing GENERATED CODE warnings so as they dont go to java doc site.
Adding documentation to setters.
Maintaining in doc JS annotations for reference